fix(aat): bring package in from tarball instead of artifactory
