+++ jest.config.js
@@ -13,7 +13,11 @@ module.exports = {
     },
   },
   setupFiles: ['<rootDir>/config/jest/setup.js'],
-  testMatch: ['<rootDir>/**/__tests__/**/*.js?(x)', '<rootDir>/**/?(*.)(spec|test).js?(x)'],
+  testMatch: [
+    '<rootDir>/**/__tests__/**/*.js?(x)',
+    '<rootDir>/**/?(*.)(spec|test).js?(x)',
+    '<rootDir>/**/?(*.)test.a11y.js?(x)',
+  ],
   testURL: 'http://localhost',
   transform: {
     '^.+\\.jsx?$': 'babel-jest',
