+++ package.json
@@ -23,7 +23,7 @@
     "start": "yarn storybook",
     "storybook": "start-storybook -p 3000 -s public/development",
     "test": "jest --no-cache",
-    "test:a11y": "jest src/a11y/__tests__/",
+    "test:a11y": "jest --testPathPattern='src\\/([^\\s]+)\\/([a-zA-Z0-9\\s_\\\\.\\-\\(\\):])+(\\.test\\.a11y\\.)+(js|jsx)$'",
     "test:watch": "jest --watchAll",
     "test:update": "jest -u",
     "test:clear": "jest --clearCache"
@@ -92,7 +92,7 @@
     "@babel/runtime": "^7.0.0",
     "@commitlint/cli": "^7.2.1",
     "@commitlint/config-conventional": "^7.1.2",
-    "@ibma/aat": "^1.3.0",
+    "@ibma/aat": "https://able.ibm.com/tools/dist/aat.tgz",
     "@semantic-release/commit-analyzer": "^6.1.0",
     "@semantic-release/git": "^7.0.7",
     "@semantic-release/npm": "^5.1.3",
